refactor(store-core): hoist viewer↔pending-write correlation helpers
pathMatchesViewer and findPendingWriteForFile were duplicated byte-for-byte
in the dashboard's and app's ViewerPreWriteReview.tsx (#6857). These are
safety-relevant: a divergence in the match logic could correlate the wrong
pending write to the file open in the viewer, approving one file's edit
believing it's another's.

Move both into packages/store-core/src/pending-permissions.ts, next to
isLivePermissionPrompt which centralizes the sibling predicate for the same
anti-drift reason (#5759). isReviewableTool/TOOL_DIFF stay local to each
client's PreWriteDiffReview.tsx (a presentation concern) and are passed into
findPendingWriteForFile as an injected predicate.

Both clients now import the helpers from @chroxy/store-core and no longer
carry local copies. Pure-logic unit tests for the two helpers moved from
each client's ViewerPreWriteReview.test.(ts|tsx) into store-core's
pending-permissions.test.ts, with added coverage for backslash-path
normalization, first-match selection among multiple candidates, injected
predicate behavior, and now/staleness gating.

Closes #6859
